Use inheritance to abstract away common methods
