+++ models/base_model.py
@@ -1,0 +1,17 @@
+from db import db
+
+
+class BaseModel(db.Model):
+    __abstract__ = True
+
+    @classmethod
+    def find_by_id(cls, id):
+        return cls.query.filter_by(id=id).first()
+
+    def save_to_db(self):
+        db.session.add(self)
+        db.session.commit()
+
+    def delete_from_db(self):
+        db.session.delete(self)
+        db.session.commit()
