+++ models/revoked_tokens.py
@@ -1,7 +1,8 @@
 from db import db
+from models.base_model import BaseModel
 
 
-class RevokedTokensModel(db.Model):
+class RevokedTokensModel(BaseModel):
     __tablename__ = 'revoked_tokens'
 
     id = db.Column(db.Integer, primary_key=True)
@@ -10,11 +11,6 @@ class RevokedTokensModel(db.Model):
     def __init__(self, jti):
         self.jti = jti
 
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-    
     @classmethod
     def is_jti_blacklisted(cls, jti):
-        query = cls.query.filter_by(jti=jti).first()
-        return bool(query)
+        return bool(cls.find_by_id(jti))
